Import SwiftAudio via podspec
