+++ react-native-track-player.podspec
@@ -18,4 +18,5 @@ Pod::Spec.new do |s|
   s.swift_version = "4.2"
 
   s.dependency "React-Core"
+  s.dependency "SwiftAudioEx", "0.13.2"
 end
